refactored logic into App, refactored SearchBar to be stateless
